Various changes to prevent narrowing conversions: chrome/
Bug: 1216696
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2957542
Auto-Submit: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Yaron Friedman <yfriedman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#892073}

+++ chrome/browser/about_flags_browsertest.cc
@@ -3,6 +3,8 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+#include "chrome/browser/about_flags.h"
+
 #include "base/command_line.h"
 #include "base/strings/stringprintf.h"
 #include "base/strings/utf_string_conversions.h"
@@ -10,7 +12,6 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 #include "base/test/scoped_feature_list.h"
 #include "build/build_config.h"
 #include "build/chromeos_buildflags.h"
-#include "chrome/browser/about_flags.h"
 #include "chrome/browser/ui/browser.h"
 #include "chrome/browser/ui/tabs/tab_strip_model.h"
 #include "chrome/browser/unexpire_flags.h"
@@ -137,14 +138,17 @@ void WaitForExperimentalFeatures(content::WebContents* contents) {
 const std::vector<flags_ui::FeatureEntry> GetFeatureEntries(
     const std::string& unexpire_name) {
   std::vector<flags_ui::FeatureEntry> entries = {
-      {kFlagName, "name-1", "description-1", -1,
+      {kFlagName, "name-1", "description-1", static_cast<unsigned short>(-1),
        ORIGIN_LIST_VALUE_TYPE(kSwitchName, "")},
-      {kExpiredFlagName, "name-2", "description-2", -1,
+      {kExpiredFlagName, "name-2", "description-2",
+       static_cast<unsigned short>(-1),
        SINGLE_VALUE_TYPE(kExpiredFlagSwitchName)},
-      {kFlagWithOptionSelectorName, "name-3", "description-3", -1,
+      {kFlagWithOptionSelectorName, "name-3", "description-3",
+       static_cast<unsigned short>(-1),
        SINGLE_VALUE_TYPE(kFlagWithOptionSelectorSwitchName)}};
   flags_ui::FeatureEntry expiry_entry = {
-      unexpire_name.c_str(), "unexpire name", "unexpire desc", -1,
+      unexpire_name.c_str(), "unexpire name", "unexpire desc",
+      static_cast<unsigned short>(-1),
       SINGLE_VALUE_TYPE("unexpire-dummy-switch")};
   entries.push_back(expiry_entry);
   return entries;
